2024/5/9 기본수정완료
css수정 및 daisy ui 도입

+++ src/main/java/com/JYProject/project/model/Board.java
@@ -40,10 +40,10 @@ public class Board {
     private int viewCnt;
 
     @Comment("좋아요")
-    private Long likes;
+    private int likes;
 
     @Comment("싫어요")
-    private Long dislikes;
+    private int dislikes;
 
     @Comment("카레고리번호")
     private int categoryId;
